Turn to use __uuidof, now wsl/winadapter.h are platform independent
