+++ googletest/feature_support_test.cpp
@@ -2,11 +2,7 @@
 // Licensed under the MIT License.
 #include "gtest/gtest.h"
 
-#ifndef _WIN32
 #include <wsl/winadapter.h>
-#elif defined(__MINGW32__)
-#include <unknwn.h>
-#endif
 
 #include <directx/d3d12.h>
 #include <directx/dxcore.h>
